+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.4-pre
 * Add buildlight command
+* Ninja command's job and schedule are now flags since they are optional
 
 ### 0.1.3
 * Fix build trigger error on Jenkins v1.5xx when job requires auth